requirements.txt
flask
psycopg[binary]
python-dotenv
redis
types-redis
